requirements.txt
Flask
Flask-SQLAlchemy
psycopg2-binary  
python-dotenv    
pytest
pytest-flask
coverage
Werkzeug